Revert "Feature/introduction"
