+++ nbc-nep-front/package.json
@@ -14,10 +14,8 @@
     "@supabase/auth-helpers-nextjs": "^0.8.7",
     "@supabase/supabase-js": "^2.39.2",
     "@tanstack/react-query": "^5.17.9",
-    "@types/lodash": "^4.14.202",
     "axios": "^1.6.5",
     "framer-motion": "^11.0.3",
-    "lodash": "^4.17.21",
     "mediasoup-client": "^3.7.2",
     "next": "14.0.4",
     "phaser": "^3.70.0",
