+++ nbc-nep-front/public/assets/introduction_2.png
