+++ nbc-nep-front/src/components/home/home.constant.ts
@@ -3,34 +3,34 @@ export const contents = [
     title: "다중 사용자 화상 • 화면 공유",
     description:
       "스페이스에 접속한 모든 유저와 화상 및 음성 • 화면을 공유하세요 ",
-    videoSrc: "/assets/video/mediashare.mp4",
+    imgSrc: "/assets/home/content-icon1.svg",
     initialY: -100,
-    top: "12%",
+    top: "10%",
     left: "35%",
   },
   {
-    title: "최대 4개 화면 공유",
+    title: "최대 4개 다중 화면 공유",
     description:
-      "최대 4개의 화면 공유를 통해 회의 간 불필요한 작업을 최소화하세요",
-    videoSrc: "/assets/video/pinch.mp4",
+      "최대 4개의 화면 공유를 통해 회의 간 불필요한 작업을 최소화 하세요",
+    imgSrc: "/assets/home/content-icon2.svg",
     initialY: -100,
-    top: "7%",
-    left: "calc(35% + 39rem)",
+    top: "5%",
+    left: "calc(35% + 35rem)",
   },
   {
-    title: "화면 배치 및 크기 조절",
+    title: "화면 배치 및 크기조절",
     description: "원하는 화면을 줌, 드래그를 통해 더 자세히 확인하세요",
-    videoSrc: "/assets/video/walk.mp4",
+    imgSrc: "/assets/home/content-icon4.svg",
     initialY: 100,
-    top: "52%",
+    top: "45%",
     left: "35%",
   },
   {
-    title: "화면 선택 및 레이아웃 생성",
+    title: "화면 선택 및 레이아웃 분할",
     description: "다른 유저의 화면을 다양한 레이아웃으로 확인하세요",
-    videoSrc: "/assets/video/mediashare.mp4",
+    imgSrc: "/assets/home/content-icon3.svg",
     initialY: 100,
-    top: "47%",
-    left: "calc(35% + 39rem)",
+    top: "40%",
+    left: "calc(35% + 35rem)",
   },
 ];
