+++ nbc-nep-front/src/components/layout/Footer.tsx
@@ -28,6 +28,7 @@ export default function Footer() {
 }
 
 const StFooter = styled.footer`
+  position: fixed;
   width: 100%;
   bottom: 0;
   left: 0%;
